# Improve resolver speed