@epage 2024-10-10
> Absolutely love the switch to rustc-hash, https://github.com/Eh2406/pubgrub-crates-benchmark/issues/6#issuecomment-2403545404 this is a huge speed up with no downside.

Should we split the test and rustc-hash commits out into its own PR and get that merged while we work through nohash?