@weihanglo 2024-10-10
While it doesn't seem like this part of code is under the risk of hash collision, there is a concern around the other part in Cargo, see <https://github.com/rust-lang/cargo/issues/13171#issuecomment-2181266802>.

I also did some exploration of blake3 as a stable hash algorithm candidate, and blake3 has become a dependency since #14137. Have we benchmarked that as well?